client/requirements.txt
CTkMessagebox==2.7
CTkTable==1.1
customtkinter==5.2.2
Pillow==11.0.0
pycryptodome==3.20.0
qrcode==8.0
argon2-cffi==23.1.0